+++ wuxin-paotui-server/src/main/java/com/wuxin/mapper/MerchantInfoMapper.java
@@ -1,0 +1,9 @@
+package com.wuxin.mapper;
+
+import com.baomidou.mybatisplus.core.mapper.BaseMapper;
+import com.wuxin.entity.MerchantInfoEntity;
+import org.apache.ibatis.annotations.Mapper;
+
+@Mapper
+public interface MerchantInfoMapper extends BaseMapper<MerchantInfoEntity> {
+}